google SSO OAuth endpoint
Added the backend side of the Google SSO OAuth flow. Once the /auth endpoint is hit, it should create a JWT from the google token and then sets that as a cookie. Also added example middleware and protected endpoint route at "protected/hi" that requires a cookie to access. Might need to change the code a bit to actually put the cookie in the frontend.